+++ rstest_macros/src/lib.rs
@@ -1,12 +1,11 @@
 #![allow(clippy::test_attr_in_doctest)]
+#![allow(unexpected_cfgs)]
 #![cfg_attr(use_proc_macro_diagnostic, feature(proc_macro_diagnostic))]
 extern crate proc_macro;
 
 // Test utility module
 #[cfg(test)]
 pub(crate) mod test;
-#[cfg(doc)]
-use rstest_reuse;
 
 #[macro_use]
 mod error;
@@ -884,6 +883,39 @@ pub fn fixture(
 /// Just the attributes that ends with `test` (last path segment) can be injected:
 /// in this case the `#[actix_rt::test]` attribute will replace the standard `#[test]`
 /// attribute.
+///
+/// ## Local timeout and `#[by_ref]` attribute
+///
+/// In some cases you may want to use a local lifetime for some arguments of your test.
+/// In these cases you can use the `#[by_ref]` attribute then use the reference instead
+/// the value.
+///
+/// ```rust
+/// # use std::cell::Cell;
+/// # use rstest::*;
+/// # #[derive(Debug, Clone, Copy, PartialEq, Eq)]
+/// enum E<'a> {
+///     A(bool),
+///     B(&'a Cell<E<'a>>),
+/// }
+///
+/// fn make_e_from_bool<'a>(_bump: &'a (), b: bool) -> E<'a> {
+///     E::A(b)
+/// }
+///
+/// #[fixture]
+/// fn bump() -> () {}
+///  
+/// #[rstest]
+/// #[case(true, E::A(true))]
+/// fn it_works<'a>(#[by_ref] bump: &'a (), #[case] b: bool, #[case] expected: E<'a>) {
+///     let actual = make_e_from_bool(&bump, b);
+///     assert_eq!(actual, expected);
+/// }
+/// ```
+///
+/// You can use `#[by_ref]` attribute for all arguments of your test and not just for fixture
+/// but also for cases, values and files.
 ///
 /// ## Putting all Together
 ///